Improved: code to use ion-menu for display filters on mobile device and removed the usage of ion-datetime and used ion-input(#22xecka)
